+++ base/src/main/java/proguard/evaluation/Variables.java
@@ -269,6 +269,19 @@ public class Variables
         return load(index).instructionOffsetValue();
     }
 
+    /**
+     * Replaces all the references to {@param toReplace} with references to {@param replacement}.
+     */
+    public void replaceReferences(Value toReplace, Value replacement)
+    {
+        for (int i = 0; i < values.length; i++)
+        {
+            if (values[i] == toReplace)
+            {
+                values[i] = replacement;
+            }
+        }
+    }
 
     // Implementations for Object.